[PATCH 0] embedshim: rewrite bundle handler
The bundle is the key store for reload. Besides the
rootfs/init.pid/config.json, the embedshim needs to store
stdio/options/eventID important information in bundle.

In order to manage files in bundle easily, this commit introduces option
design as newBundle' interface and bring helper to read the file in
bundle.

Signed-off-by: Wei Fu <fuweid89@gmail.com>