[PATCH 2] embedshim: rewrite init process
Basically, we should reuse the upstream pkg/process package. But the IO
design doesn't work with that. The initProcess should be redesigned to
work with embedshim.

In this commit, we rename Init to initProcess because we don't need to
export it. And then we make newInitProcess with bundle since the bundle
is the key store dir for us, especially when containerd restart.

Signed-off-by: Wei Fu <fuweid89@gmail.com>