[PATCH 4] embedshim: rewrite pid monitor
Rewrited the id allocator because we don't have to allocate id with
namespace and task ID and release it. The uin64 number range is totally
enough for us. Just keep it simpler with nextID() interface. And the
trace event ID can also used for exec process in the future.

Added traceEventId field in the initProcess because the trace event ID
is the identity of process.

And rewrite the monitor interface:

* subscribe -> traceInitProcess
* resubscribe -> repollingInitProces

I think it can be easy to understand.

Signed-off-by: Wei Fu <fuweid89@gmail.com>