[PATCH 10] embedshim: store id and pin bpf in root dir
The host can be restarted. If we should store id allocator db in tmpfs,
it doesn't bring issue but the containerd log will show id reuse. I
think the uint64 range is bigger enough and it will be easy to debug the
trace ID when restart containerd if we store it in root dir.

BPF pinned dir is just aligned with id db.

Signed-off-by: Wei Fu <fuweid89@gmail.com>